Adding 3 base cases
